+++ src/sandboxes/olympia-realtime-light-test/olympia-terrain/OlympiaTerrain.ts
@@ -1,31 +1,33 @@
-import { DoubleSide, Group, Mesh, MeshBasicMaterial, MeshStandardMaterial, PlaneBufferGeometry, RepeatWrapping, ShaderMaterial, sRGBEncoding, Texture, TextureLoader, UniformsLib, UniformsUtils } from "three"
+import { 
+    BufferAttribute,
+    Color,
+    DataTexture,
+    DoubleSide,
+    FrontSide,
+    Group,
+    LinearEncoding,
+    Material,
+    Mesh,
+    MeshBasicMaterial,
+    MeshStandardMaterial,
+    PlaneBufferGeometry,
+    RepeatWrapping,
+    Shader,
+    ShaderChunk,
+    ShaderLib,
+    ShaderMaterial,
+    sRGBEncoding,
+    Texture,
+    TextureLoader,
+    Uniform,
+    UniformsLib,
+    UniformsUtils,
+    WebGLRenderer 
+} from "three"
 
-// export const TerrainMapLow = {
-//     heightMap: `${baseUrl}/Terrain-LowDetail/terrain_GC_export_flipy_2048.jpg`,
-//     heightScale: 117.0,
-//     textureMaps: {
-//       beachA: `${baseUrl}/Terrain-LowDetail/Beach_A.jpg`,
-//       grassB: `${baseUrl}/Terrain-LowDetail/Grass_B.jpg`,
-//       grassC: `${baseUrl}/Terrain-LowDetail/Grass_C.jpg`,
-//       gravelA: `${baseUrl}/Terrain-LowDetail/Gravel_A.jpg`,
-//       rockA: `${baseUrl}/Terrain-LowDetail/Rock_A.jpg`,
-//       soilA: `${baseUrl}/Terrain-LowDetail/Soil_A.jpg`,
-//       soilB: `${baseUrl}/Terrain-LowDetail/Soil_B.jpg`,
-//     },
-//     normalMaps: {
-//       beachA: `${baseUrl}/Terrain-LowDetail/Beach_A_norm.jpg`,
-//       grassB: `${baseUrl}/Terrain-LowDetail/Grass_B_norm.jpg`,
-//       grassC: `${baseUrl}/Terrain-LowDetail/Grass_C_norm.jpg`,
-//       gravelA: `${baseUrl}/Terrain-LowDetail/Gravel_A_normal.jpg`,
-//       rockA: `${baseUrl}/Terrain-LowDetail/Rock_A_norm.jpg`,
-//       soilA: `${baseUrl}/Terrain-LowDetail/Soil_A_norm.jpg`,
-//       soilB: `${baseUrl}/Terrain-LowDetail/Soil_B_norm.jpg`,
-//     },
-//     splatterMaps: {
-//       map1: `${baseUrl}/Terrain-Splattermap/TerrainSplattermap1.png`,
-//       map2: `${baseUrl}/Terrain-Splattermap/TerrainSplattermap2.png`,
-//     },
-//   }
+import white_tex from './textures/white_16x16.png';
+import black_tex from './textures/black_16x16.png';
+import grey_tex from './textures/grey_16x16.png';
 
 // Splatter maps.
 import splatter_map_1 from './textures/TerrainSplattermap1.png';
@@ -47,17 +49,19 @@ import soilA_tex from './textures/Soil_A.jpg';
 import soilA_norm from './textures/Soil_A_norm.jpg';
 import soilB_tex from './textures/Soil_B.jpg';
 import soilB_norm from './textures/Soil_B_norm.jpg';
+import { replace } from "lodash";
+import { stringInsertAt } from "../../../utils/MiscUtils";
 
 
 const vertexShader = `
 uniform sampler2D bumpTexture;
 uniform float bumpScale;
 
-varying vec2 vUV;
+varying vec2 vUv;
 
 void main()
 {
-    vUV = uv;
+    vUv = uv;
 
     vec4 bumpData = texture2D(bumpTexture, uv);
     float vAmount = bumpData.r;
@@ -82,12 +86,12 @@ uniform sampler2D soilBTexture;
 uniform sampler2D splatterMap1;
 uniform sampler2D splatterMap2;
 
-varying vec2 vUV;
+varying vec2 vUv;
 
 void main()
 {
-    vec4 splat1 = texture2D(splatterMap1, vUV);
-    vec4 splat2 = texture2D(splatterMap2, vUV);
+    vec4 splat1 = texture2D(splatterMap1, vUv);
+    vec4 splat2 = texture2D(splatterMap2, vUv);
 
     // the result of guesswork after decomposing the two splatmaps into RGBA
     // https://docs.gimp.org/en/plug-in-decompose-registered.html
@@ -109,18 +113,78 @@ void main()
     float soilARepeat = 311.0;
     float soilBRepeat = 667.0;
 
-    vec4 beachA = beachAAmount * texture2D(beachATexture, vUV * beachARepeat);
-    vec4 grassB = grassBAmount * texture2D(grassBTexture, vUV * grassBRepeat);
-    vec4 grassC1 = grassCAmount1 * texture2D(grassCTexture1, vUV * grassC1Repeat);
-    vec4 grassC2 = grassCAmount2 * texture2D(grassCTexture2, vUV * grassC2Repeat);
-    vec4 gravelA = gravelAAmount * texture2D(gravelATexture, vUV * gravelARepeat);
-    vec4 rockA = rockAAmount * texture2D(rockATexture, vUV * rockARepeat);
-    vec4 soilA = soilAAmount * texture2D(soilATexture, vUV * soilARepeat);
-    vec4 soilB = soilBAmount * texture2D(soilBTexture, vUV * soilBRepeat);
+    vec4 beachA = beachAAmount * texture2D(beachATexture, vUv * beachARepeat);
+    vec4 grassB = grassBAmount * texture2D(grassBTexture, vUv * grassBRepeat);
+    vec4 grassC1 = grassCAmount1 * texture2D(grassCTexture1, vUv * grassC1Repeat);
+    vec4 grassC2 = grassCAmount2 * texture2D(grassCTexture2, vUv * grassC2Repeat);
+    vec4 gravelA = gravelAAmount * texture2D(gravelATexture, vUv * gravelARepeat);
+    vec4 rockA = rockAAmount * texture2D(rockATexture, vUv * rockARepeat);
+    vec4 soilA = soilAAmount * texture2D(soilATexture, vUv * soilARepeat);
+    vec4 soilB = soilBAmount * texture2D(soilBTexture, vUv * soilBRepeat);
 
     gl_FragColor = vec4(0.0, 0.0, 0.0, 1.0) + beachA + grassB + grassC1 + grassC2 + gravelA + rockA + soilA + soilB;
 }
 `
+
+const splat_pars_fragment_glsl = `
+uniform sampler2D beachATexture;
+uniform sampler2D grassBTexture;
+uniform sampler2D grassCTexture1;
+uniform sampler2D grassCTexture2;
+uniform sampler2D gravelATexture;
+uniform sampler2D rockATexture;
+uniform sampler2D soilATexture;
+uniform sampler2D soilBTexture;
+
+uniform sampler2D splatterMap1;
+uniform sampler2D splatterMap2;
+`;
+
+const splat_fragment_glsl = `
+
+    vec4 splat1 = texture2D(splatterMap1, vUv);
+    vec4 splat2 = texture2D(splatterMap2, vUv);
+
+    // the result of guesswork after decomposing the two splatmaps into RGBA
+    // https://docs.gimp.org/en/plug-in-decompose-registered.html
+    float gravelAAmount = splat2.a;
+    float beachAAmount = splat1.b;
+    float grassCAmount1 = splat1.g;
+    float grassCAmount2 = splat1.a;
+    float soilBAmount = splat2.g;
+    float soilAAmount = splat1.r;
+    float grassBAmount = splat2.r;
+    float rockAAmount = splat2.b;
+
+    float beachARepeat = 451.0;
+    float grassBRepeat = 331.0;
+    float grassC1Repeat = 307.0;
+    float grassC2Repeat = 50.0;
+    float gravelARepeat = 97.0;
+    float rockARepeat = 451.0;
+    float soilARepeat = 311.0;
+    float soilBRepeat = 667.0;
+
+    vec4 beachA = beachAAmount * texture2D(beachATexture, vUv * beachARepeat);
+    vec4 grassB = grassBAmount * texture2D(grassBTexture, vUv * grassBRepeat);
+    vec4 grassC1 = grassCAmount1 * texture2D(grassCTexture1, vUv * grassC1Repeat);
+    vec4 grassC2 = grassCAmount2 * texture2D(grassCTexture2, vUv * grassC2Repeat);
+    vec4 gravelA = gravelAAmount * texture2D(gravelATexture, vUv * gravelARepeat);
+    vec4 rockA = rockAAmount * texture2D(rockATexture, vUv * rockARepeat);
+    vec4 soilA = soilAAmount * texture2D(soilATexture, vUv * soilARepeat);
+    vec4 soilB = soilBAmount * texture2D(soilBTexture, vUv * soilBRepeat);
+
+    // gl_FragColor = vec4(0.0, 0.0, 0.0, 1.0) + beachA + grassB + grassC1 + grassC2 + gravelA + rockA + soilA + soilB;
+    diffuseColor = vec4(0.0, 0.0, 0.0, 1.0) + beachA + grassB + grassC1 + grassC2 + gravelA + rockA + soilA + soilB;
+`;
+
+const test_fragment_glsl = `
+    diffuseColor *= vec4(1.0, 0.0, 0.0, 1.0);
+`;
+
+ShaderChunk['splat_fragment_glsl'] = splat_fragment_glsl;
+// ShaderChunk['splat_fragment_glsl'] = test_fragment_glsl;
+ShaderChunk['splat_pars_fragment_glsl'] = splat_pars_fragment_glsl;
 
 async function loadTexture(url: string, postProcess?: (texture: Texture) => void): Promise<Texture> {
     return new Promise<Texture>((resolve, reject) => {
@@ -146,68 +210,157 @@ async function loadTexture(url: string, postProcess?: (texture: Texture) => void
     });   
 }
 
+/**    
+ * Extending shaders in three really sucks.
+ * The WebGLProgram define injection system appears to read 
+ * from the material instead of from uniforms. Assigning uniform and property togethor
+ * seems to be the only way to fix issses with the WebGLProgram injections.
+ */
+function assignUniformAndProperty(material: ShaderMaterial, uniformName: string, value: any): void {
+    material.uniforms[uniformName].value = value;
+    (material as any)[uniformName] = value;
+}
+
 export async function createOlympiaTerrain(): Promise<Group> {
     const terrainGroup = new Group();
     terrainGroup.name = 'Terrain';
 
     const heightScale = 117.0;
 
+    const whiteTexture = await loadTexture(white_tex);
+
     const heightTexture = await loadTexture(height_map, (texture) => {
         texture.encoding = sRGBEncoding;
         texture.anisotropy = 16;
+        texture.needsUpdate = true;
     });
 
-    // const postProcessTexture = (texture: Texture) => {
-    //     texture.wrapS = RepeatWrapping;
-    //     texture.wrapT = RepeatWrapping;
-    //     texture.encoding = sRGBEncoding;
-    //     texture.anisotropy = 16;
-    // }
+    const postProcessTexture = (texture: Texture) => {
+        texture.wrapS = RepeatWrapping;
+        texture.wrapT = RepeatWrapping;
+        texture.encoding = sRGBEncoding;
+        texture.anisotropy = 8;
+    }
 
-    // const beachATexture = await loadTexture(beachA_tex, postProcessTexture);
-    // const grassBTexture = await loadTexture(grassB_tex, postProcessTexture);
-    // const grassCTexture1 = await loadTexture(grassC_tex, postProcessTexture);
-    // const grassCTexture2 = await loadTexture(grassC_tex, postProcessTexture);
-    // const gravelATexture = await loadTexture(gravelA_tex, postProcessTexture);
-    // const rockATexture = await loadTexture(rockA_tex, postProcessTexture);
-    // const soilATexture = await loadTexture(soilA_tex, postProcessTexture);
-    // const soilBTexture = await loadTexture(soilB_tex, postProcessTexture);
+    const beachATexture = await loadTexture(beachA_tex, postProcessTexture);
+    const grassBTexture = await loadTexture(grassB_tex, postProcessTexture);
+    const grassCTexture1 = await loadTexture(grassC_tex, postProcessTexture);
+    const grassCTexture2 = await loadTexture(grassC_tex, postProcessTexture);
+    const gravelATexture = await loadTexture(gravelA_tex, postProcessTexture);
+    const rockATexture = await loadTexture(rockA_tex, postProcessTexture);
+    const soilATexture = await loadTexture(soilA_tex, postProcessTexture);
+    const soilBTexture = await loadTexture(soilB_tex, postProcessTexture);
 
-    // const splatterMap1 = await loadTexture(splatter_map_1);
-    // const splatterMap2 = await loadTexture(splatter_map_2);
+    const splatterMap1 = await loadTexture(splatter_map_1);
+    const splatterMap2 = await loadTexture(splatter_map_2);
 
     const planeGeo = new PlaneBufferGeometry(2048, 2048, 256, 256);
-    const planeMat = new MeshStandardMaterial({
-        map: heightTexture,
-        displacementMap: heightTexture,
-        displacementScale: heightScale,
-        envMapIntensity: 1,
-        color: '#fff',
-        roughness: 1,
-    });
-    /*
+
+    // Assing uv2 attribute with value of uv. Need these for aoMap.
+    planeGeo.attributes.uv2 = planeGeo.attributes.uv;
+
+    // Create a new shader material that inherits from the mesh standard material.
+    let vertexShader = ShaderLib.standard.vertexShader;
+    let fragmentShader = ShaderLib.standard.fragmentShader;
+    let uniforms = UniformsUtils.clone(ShaderLib.standard.uniforms);
+
+    // Assing splatting textures to the material's uniforms.
+    uniforms['bumpTexture'] = { value: heightTexture };
+    uniforms['bumpScale'] = { value: heightScale };
+    uniforms['beachATexture'] = { value: beachATexture };
+    uniforms['grassBTexture'] = { value: grassBTexture };
+    uniforms['grassCTexture1'] = { value: grassCTexture1 };
+    uniforms['grassCTexture2'] = { value: grassCTexture2 };
+    uniforms['gravelATexture'] = { value: gravelATexture };
+    uniforms['rockATexture'] = { value: rockATexture };
+    uniforms['soilATexture'] = { value: soilATexture };
+    uniforms['soilBTexture'] = { value: soilBTexture };
+    uniforms['splatterMap1'] = { value: splatterMap1 };
+    uniforms['splatterMap2'] = { value: splatterMap2 };
+
+    // Insert the splat glsl uniform declartions.
+    fragmentShader = splat_pars_fragment_glsl + '\n' + fragmentShader;
+
+    // Insert the splat fragment right after the map fragment.
+    fragmentShader = fragmentShader.replace(
+        '#include <map_fragment>',
+        '#include <splat_fragment_glsl>',
+    );
+    // const mapFragInclude = '#include <map_fragment>';
+    // const mapFragStart = fragmentShader.indexOf(mapFragInclude);
+    // fragmentShader = stringInsertAt(fragmentShader, '\n#include <splat_fragment_glsl>', mapFragStart + mapFragInclude.length);
+
+    // console.log(`terrain fragment shader before compile:`, fragmentShader);
+    
     const planeMat = new ShaderMaterial({
-        uniforms: {
-            ...lightUniforms,
-            bumpTexture: { value: heightTexture },
-            bumpScale: { value: heightScale },
-            beachATexture: { value: beachATexture },
-            grassBTexture: { value: grassBTexture },
-            grassCTexture1: { value: grassCTexture1 },
-            grassCTexture2: { value: grassCTexture2 },
-            gravelATexture: { value: gravelATexture },
-            rockATexture: { value: rockATexture },
-            soilATexture: { value: soilATexture },
-            soilBTexture: { value: soilBTexture },
-            splatterMap1: { value: splatterMap1 },
-            splatterMap2: { value: splatterMap2 },
-        },
+        uniforms,
         vertexShader,
         fragmentShader,
-        side: DoubleSide,
-        lights: true
+        side: FrontSide,
+        lights: true,
+        name: 'TerrainMaterial',
     });
-    */
+
+    (window as any).planeMat = planeMat;
+    
+    (planeMat as any).isMeshStandardMaterial = true;
+    
+    // Assign these uniforms and properties after the material is created.
+    // This gets around three's handling of ShaderMaterial differently from 
+    // the MeshStandardMaterial when compiling it in WebGLProgram.
+
+    // Assign map property so that shader defines for uvs get injected by WebGLProgram.
+    // The texture is otherwise unused as it gets overwritten by the splatter.
+    assignUniformAndProperty(planeMat, 'map', whiteTexture);
+
+    // Use standard shader's displacement functionality for terrain.
+    assignUniformAndProperty(planeMat, 'displacementMap', heightTexture);
+    assignUniformAndProperty(planeMat, 'displacementScale', heightScale);
+    
+    assignUniformAndProperty(planeMat, 'envMapIntensity', 1);
+    assignUniformAndProperty(planeMat, 'metalness', 0);
+    assignUniformAndProperty(planeMat, 'roughness', 1);
+
+    assignUniformAndProperty(planeMat, 'aoMap', heightTexture);
+    assignUniformAndProperty(planeMat, 'aoMapIntensity', 1);
+
+
+    // const planeMat = new MeshStandardMaterial({
+    //     displacementMap: heightTexture,
+    //     displacementScale: heightScale,
+    //     aoMapIntensity: 5,
+    //     aoMap: heightTexture,
+    //     envMapIntensity: 1,
+    //     color: '#fff',
+    //     roughness: 1,
+    // });
+
+    // planeMat.onBeforeCompile = (shader: Shader, renderer: WebGLRenderer) => {
+    // };
+
+    // const lightUniforms = UniformsUtils.merge([
+    //     UniformsLib.lights
+    // ]);
+
+    // const planeMat = new ShaderMaterial({
+    //     uniforms: {
+    //         bumpTexture: { value: heightTexture },
+    //         bumpScale: { value: heightScale },
+    //         beachATexture: { value: beachATexture },
+    //         grassBTexture: { value: grassBTexture },
+    //         grassCTexture1: { value: grassCTexture1 },
+    //         grassCTexture2: { value: grassCTexture2 },
+    //         gravelATexture: { value: gravelATexture },
+    //         rockATexture: { value: rockATexture },
+    //         soilATexture: { value: soilATexture },
+    //         soilBTexture: { value: soilBTexture },
+    //         splatterMap1: { value: splatterMap1 },
+    //         splatterMap2: { value: splatterMap2 },
+    //     },
+    //     vertexShader,
+    //     fragmentShader,
+    //     side: DoubleSide,
+    // });
 
     const planeMesh = new Mesh(planeGeo, planeMat);
     planeMesh.rotation.x = -Math.PI / 2;